web_application/requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-custom-user==0.7
django-registration==2.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
numpy==1.13.3
pandas==0.20.3
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
vincent==0.4.4
